requirements.txt
aiogram>=3.0.0
ccxt>=4.0.0
aiosqlite
pandas
numpy
python-dotenv